# openapi-3.0.yaml
openapi: 3.0.3
info:
  title: BNA REST API
  description: Provides a way to retrieve the BNA results programmatically.
  contact:
    name: The BNA Mechanics team
    url: https://peopleforbikes.github.io/
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development API
  - url: https://api.peopleforbikes.xyz
    description: Staging API
paths:
...
  /ratings:
    get:
      tags:
        - rating
      description: Get city ratings
      operationId: get_ratings
      parameters:
        - name: page_size
          in: query
          description: The number of items per page
          required: false
          schema:
            type: integer
            format: int64
            maximum: 65536
            minimum: 1
          example: 25
        - name: page
          in: query
          description: The result page being returned
          required: false
          schema:
            type: integer
            format: int64
            maximum: 65536
            minimum: 1
          example: 5
      responses:
        "200":
          description: Fetches the city ratings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ratings"
    post:
      tags:
        - rating
      description: Create a new city rating
      operationId: post_rating
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingPost"
        required: true
      responses:
        "201":
          description: Creates a new city rating
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
        "400":
          description: >-
            The request was formatted incorrectly or missing required
            parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      the request was formatted incorrectly or missing required
                      parameters
                    id: blfwkg8nvHcEJnQ=
                    source:
                      parameter: status
                    status: "400"
                    title: Bad Request
        "401":
          description: >-
            The request has not been fulfilled because it lacks valid
            authentication credentials for the target resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      invalid authentication credentials to access the specified
                      resource
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "401"
                    title: Unauthorized
        "403":
          description: >-
            Forbidden to make the request. Most likely this indicates an issue
            with the credentials or permissions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: access to the requested resource is forbidden
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "403"
                    title: Forbidden
        "404":
          description: >-
            The particular resource requested was not found. This occurs, for
            example, when the id of the requested  resource does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: the resource was not found
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "404"
                    title: Item Not Found
  /ratings/{rating_id}:
    get:
      tags:
        - rating
      description: Get the details of a specific city rating
      operationId: get_rating
      parameters:
        - name: rating_id
          in: path
          description: Rating identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Fetches the details of a city rating
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
        "400":
          description: >-
            The request was formatted incorrectly or missing required
            parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      the request was formatted incorrectly or missing required
                      parameters
                    id: blfwkg8nvHcEJnQ=
                    source:
                      parameter: status
                    status: "400"
                    title: Bad Request
        "401":
          description: >-
            The request has not been fulfilled because it lacks valid
            authentication credentials for the target resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      invalid authentication credentials to access the specified
                      resource
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "401"
                    title: Unauthorized
        "403":
          description: >-
            Forbidden to make the request. Most likely this indicates an issue
            with the credentials or permissions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: access to the requested resource is forbidden
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "403"
                    title: Forbidden
        "404":
          description: >-
            The particular resource requested was not found. This occurs, for
            example, when the id of the requested  resource does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: the resource was not found
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "404"
                    title: Item Not Found
  /ratings/{rating_id}/city:
    get:
      tags:
        - rating
      description: Get a city rating and its associated city details
      operationId: get_ratings_city
      parameters:
        - name: rating_id
          in: path
          description: Rating identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Fetches a city rating and its associated city details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingWithCity"
        "400":
          description: >-
            The request was formatted incorrectly or missing required
            parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      the request was formatted incorrectly or missing required
                      parameters
                    id: blfwkg8nvHcEJnQ=
                    source:
                      parameter: status
                    status: "400"
                    title: Bad Request
        "401":
          description: >-
            The request has not been fulfilled because it lacks valid
            authentication credentials for the target resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      invalid authentication credentials to access the specified
                      resource
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "401"
                    title: Unauthorized
        "403":
          description: >-
            Forbidden to make the request. Most likely this indicates an issue
            with the credentials or permissions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: access to the requested resource is forbidden
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "403"
                    title: Forbidden
        "404":
          description: >-
            The particular resource requested was not found. This occurs, for
            example, when the id of the requested  resource does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: the resource was not found
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "404"
                    title: Item Not Found
...
  /usstates:
    get:
      tags:
        - usstate
      description: Get the details of all US states.
      operationId: get_us_states
      parameters:
        - name: page_size
          in: query
          description: The number of items per page
          required: false
          schema:
            type: integer
            format: int64
            maximum: 65536
            minimum: 1
          example: 25
        - name: page
          in: query
          description: The result page being returned
          required: false
          schema:
            type: integer
            format: int64
            maximum: 65536
            minimum: 1
          example: 5
      responses:
        "200":
          description: Fetches US states
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsStates"
  /usstates/{name}:
    get:
      tags:
        - usstate
      description: Get the details of a specific US state.
      operationId: get_us_state
      parameters:
        - name: name
          in: path
          description: Full name of a US state
          required: true
          schema:
            type: string
          example: Texas
      responses:
        "200":
          description: Fetches a Us state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsState"
        "400":
          description: >-
            The request was formatted incorrectly or missing required
            parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      the request was formatted incorrectly or missing required
                      parameters
                    id: blfwkg8nvHcEJnQ=
                    source:
                      parameter: status
                    status: "400"
                    title: Bad Request
        "401":
          description: >-
            The request has not been fulfilled because it lacks valid
            authentication credentials for the target resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      invalid authentication credentials to access the specified
                      resource
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "401"
                    title: Unauthorized
        "403":
          description: >-
            Forbidden to make the request. Most likely this indicates an issue
            with the credentials or permissions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: access to the requested resource is forbidden
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "403"
                    title: Forbidden
        "404":
          description: >-
            The particular resource requested was not found. This occurs, for
            example, when the id of the requested  resource does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: the resource was not found
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "404"
                    title: Item Not Found
components:
  schemas:
    APIError:
      type: object
      description: >-
        Single API Error object as described in
        <https://jsonapi.org/format/#error-objects>.
      required:
        - status
        - title
        - details
        - source
      properties:
        details:
          type: string
          description: >-
            A human-readable explanation specific to this occurrence of the
            problem
          example: the entry was not found
        id:
          type: string
          description: A unique identifier for this particular occurrence of the problem.
          example: blfwkg8nvHcEJnQ=
          nullable: true
        source:
          $ref: "#/components/schemas/APIErrorSource"
          description: >-
            An object containing references to the primary source of the error.


            This field may be omitted in some situation. For instance, if the
            server cannot

            parse the request as valid JSON, including source doesn’t make sense

            (because there’s no JSON document for source to refer to).
        status:
          type: string
          description: >-
            The HTTP status code applicable to this problem, expressed as a
            string value.
          example: "404"
        title:
          type: string
          description: A short, human-readable summary of the problem
          example: Item Not Found
...
    APIErrors:
      type: object
      description: >-
        Error objects MUST be returned as an array keyed by errors in the top
        level of a

        JSON:API document.
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/APIError"
...
    City:
      type: object
      description: Detailed information of a city
      required:
        - id
        - country
        - state
        - name
        - created_at
      properties:
        country:
          $ref: "#/components/schemas/Country"
          description: Country name
        created_at:
          type: string
          format: date-time
          description: Creation date
        id:
          type: string
          format: uuid
          description: City identifier
        latitude:
          type: number
          format: double
          description: >-
            Geographic coordinate that specifies the north-south position of a
            point

            on the surface of the Earth.
          example: "51.260197"
          nullable: true
        longitude:
          type: number
          format: double
          description: >-
            Geographic coordinate that specifies the east–west position of a
            point

            on the surface of the Earth.
          example: "4.402771"
          nullable: true
        name:
          type: string
          description: City name
          example: Antwerp
        region:
          type: string
          description: >-
            Region name. A region can be a state, a province, a community, or

            something similar depending on the country. If a country does not
            have

            this concept, then the country name is used.
          example: Antwerp
          nullable: true
        residential_speed_limit:
          type: integer
          format: int32
          description: |-
            Residential speed limit in kilometer per hour (km/h).
            Only use if different from the state speed limit.
          example: "50"
          nullable: true
        state:
          type: string
          description: State name
          example: Antwerp
        state_abbrev:
          type: string
          description: A short version of the state name, usually 2 or 3 character long
          example: VAN
          nullable: true
        updated_at:
          type: string
          format: date-time
          description: Update date
          nullable: true
...
    CoreServices:
      type: object
      properties:
        dentists:
          type: number
          format: double
          description: BNA category subscore for access to dentists.
          maximum: 100
          minimum: 0
          nullable: true
        doctors:
          type: number
          format: double
          description: BNA category subscore for access to doctors.
          maximum: 100
          minimum: 0
          nullable: true
        grocery:
          type: number
          format: double
          description: BNA category subscore for access to grocery stores.
          maximum: 100
          minimum: 0
          nullable: true
        hospitals:
          type: number
          format: double
          description: BNA category subscore for access to hospitals.
          maximum: 100
          minimum: 0
          nullable: true
        pharmacies:
          type: number
          format: double
          description: BNA category subscore for access to pharmacies.
          maximum: 100
          minimum: 0
          nullable: true
        score:
          type: number
          format: double
          description: BNA category score for access to core services.
          maximum: 100
          minimum: 0
          nullable: true
        social_services:
          type: number
          format: double
          description: BNA category subscore for access to social services.
          maximum: 100
          minimum: 0
          nullable: true
    Country:
      type: string
      enum:
        - Australia
        - Belgium
        - Brazil
        - Canada
        - Chile
        - Colombia
        - Croatia
        - Cuba
        - England
        - France
        - Germany
        - Greece
        - Guatemala
        - Iran
        - Iraq
        - Ireland
        - Italy
        - Mexico
        - Netherlands
        - New Zealand
        - Northern Ireland
        - Portugal
        - Scotland
        - Spain
        - United States
        - Vietnam
        - Wales
...
    Infrastructure:
      type: object
      properties:
        high_stress_miles:
          type: number
          format: double
          description: Total miles of high-stress streets in the measured area.
          example: 253
          nullable: true
        low_stress_miles:
          type: number
          format: double
          description: Total miles of low-stress streets and paths in the measured area.
          example: 127
          nullable: true
    Measure:
      type: object
      properties:
        buffered_lane:
          type: number
          format: double
          description: Miles of buffered bike lanes.
          example: 53.859
          nullable: true
        lane:
          type: number
          format: double
          description: Miles of bike lanes.
          example: 0
          nullable: true
        path:
          type: number
          format: double
          description: Miles of off-street paths.
          example: 53.859
          nullable: true
        sharrow:
          type: number
          format: double
          description: Miles of sharrows.
          example: 53.859
          nullable: true
        track:
          type: number
          format: double
          description: Miles of tracks.
          example: 53.859
          nullable: true
    Opportunity:
      type: object
      properties:
        employment:
          type: number
          format: double
          description: BNA category subscore for access to job location areas.
          maximum: 100
          minimum: 0
          nullable: true
        higher_education:
          type: number
          format: double
          description: BNA category subscore for access to universities and colleges.
          maximum: 100
          minimum: 0
          nullable: true
        k12_education:
          type: number
          format: double
          description: BNA category subscore for access to k12 schools
          maximum: 100
          minimum: 0
          nullable: true
        score:
          type: number
          format: double
          description: BNA category score for access to education and jobs.
          maximum: 100
          minimum: 0
          nullable: true
        technical_vocational_college:
          type: number
          format: double
          description: >-
            BNA category subscore for access to technical and vocational
            colleges.
          maximum: 100
          minimum: 0
          nullable: true
    People:
      type: object
      properties:
        people:
          type: number
          format: double
          description: BNA category score for access to residential areas.
          maximum: 100
          minimum: 0
          nullable: true
...
    Rating:
      type: object
      required:
        - id
        - city_id
        - score
        - version
        - infrastructure
        - recreation
        - opportunity
        - core_services
        - people
        - retail
        - transit
        - measure
      properties:
        city_id:
          type: string
          format: uuid
          description: City identifier
        core_services:
          $ref: "#/components/schemas/CoreServices"
          description: BNA Core Services
        id:
          type: string
          format: uuid
          description: Rating identifier
        infrastructure:
          $ref: "#/components/schemas/Infrastructure"
          description: BNAInfrastructure
        measure:
          $ref: "#/components/schemas/Measure"
          description: Measurement of the city's bike infrastructure.
        opportunity:
          $ref: "#/components/schemas/Opportunity"
          description: BNA Opportunity
        people:
          $ref: "#/components/schemas/People"
          description: BNA People
        recreation:
          $ref: "#/components/schemas/Recreation"
          description: BNA Recreation
        retail:
          $ref: "#/components/schemas/Retail"
          description: BNA Retail
        score:
          type: number
          format: double
          description: Total rating score
          maximum: 100
          minimum: 0
        transit:
          $ref: "#/components/schemas/Transit"
          description: BNA Transit
        version:
          type: string
          description: >-
            Rating version

            The format follows the [calver](https://calver.org) specification
            with

            the YY.0M[.Minor] scheme.
    RatingPost:
      type: object
      required:
        - city_id
        - version
        - score
        - pop_size
        - population
        - infrastructure
        - recreation
        - opportunity
        - core_services
        - people
        - retail
        - transit
        - measure
      properties:
        city_id:
          type: string
          format: uuid
          description: City identifier
        core_services:
          $ref: "#/components/schemas/CoreServices"
          description: BNA Core Services
        infrastructure:
          $ref: "#/components/schemas/Infrastructure"
          description: BNAInfrastructure
        measure:
          $ref: "#/components/schemas/Measure"
          description: Measurement of the city's bike infrastructure.
        opportunity:
          $ref: "#/components/schemas/Opportunity"
          description: BNA Opportunity
        people:
          $ref: "#/components/schemas/People"
          description: BNA People
        pop_size:
          type: integer
          format: int32
          description: City population size category (small, medium, large).
        population:
          type: integer
          format: int32
          description: >-
            City population based on the annual U.S. Census American Community
            Survey.
        recreation:
          $ref: "#/components/schemas/Recreation"
          description: BNA Recreation
        retail:
          $ref: "#/components/schemas/Retail"
          description: BNA Retail
        score:
          type: number
          format: double
          description: City rating score
        speed_limit_override:
          type: integer
          format: int32
          description: Residential speed limit, if any.
          nullable: true
        transit:
          $ref: "#/components/schemas/Transit"
          description: BNA Transit
        version:
          type: string
          description: >-
            Rating version

            The format follows the [calver](https://calver.org) specification
            with

            the YY.0M[.Minor] scheme.
...
    RatingWithCity:
      type: object
      required:
        - rating
        - city
      properties:
        city:
          $ref: "#/components/schemas/City"
        rating:
          $ref: "#/components/schemas/Rating"
    Ratings:
      type: array
      items:
        $ref: "#/components/schemas/Rating"
    Recreation:
      type: object
      properties:
        community_centers:
          type: number
          format: double
          description: BNA category subscore for access to community centers.
          maximum: 100
          minimum: 0
          nullable: true
        parks:
          type: number
          format: double
          description: BNA category subscore for access to parks.
          maximum: 100
          minimum: 0
          nullable: true
        score:
          type: number
          format: double
          description: BNA category score for access to recreational facilities.
          maximum: 100
          minimum: 0
          nullable: true
        trails:
          type: number
          format: double
          description: BNA category subscore for access to bikeable trails.
          maximum: 100
          minimum: 0
          nullable: true
    Retail:
      type: object
      properties:
        retail:
          type: number
          format: double
          description: BNA category score for access to major retail centers.
          maximum: 100
          minimum: 0
          nullable: true
...
    Transit:
      type: object
      properties:
        transit:
          type: number
          format: double
          description: BNA category score for access to major transit stops.
          maximum: 100
          minimum: 0
          nullable: true
    UsState:
      type: object
      description: Detailed information of a city
      required:
        - name
        - abbrev
        - fipscode
        - speed_limit
      properties:
        abbrev:
          type: string
          description: Two-letter state abbreviation..
          example: TX
        fipscode:
          type: string
          description: State FIPS code.
          example: "48"
        name:
          type: string
          description: State name.
          example: Texas
        speed_limit:
          type: integer
          format: int32
          description: State speed limit in mph.
          example: "30"
    UsStates:
      type: array
      items:
        $ref: "#/components/schemas/UsState"
...